requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.23.0
Pillow>=8.0.0
opencv-python>=4.4.0
tqdm>=4.50.0